phinterval: Set Operations on Time Intervals

Implements the phinterval vector class for representing time spans that may contain gaps (disjoint intervals) or be empty. This class generalizes the 'lubridate' package's interval class to support vectorized set operations (intersection, union, difference, complement) that always return a valid time span, even when disjoint or empty intervals are created.

See also

Author

Maintainer: Ethan Sansom ethan.sansom29@gmail.com (ORCID) [copyright holder]